services/stream-processor-service/requirements.txt
kafka-python==2.0.2
python-snappy==0.7.1